fix: address all 29 PR #106 review comments
- LazyNode: fix variadic input duplication (overflow stored full array,
  causing doubled first 3 inputs in RealizeInputs/GetInputNodes)
- LazyNode: roll back IsRealized on exception so stale data isn't served
- CompiledTrainingPlan: seed loss gradient by tensor identity not array
  position (HashSet iteration order is non-deterministic)
- CompiledTrainingPlan: check TryGemmEx return values with engine
  fallback when BLAS unavailable
- CompiledTrainingPlan: use accumulation (TensorAddInPlace) instead of
  CopyTo for Add/Sub/Mul backward to handle multi-consumer tensors
- ActivationHandler: fix backward API to pass pre-activation input;
  Sigmoid/Tanh/Softmax handlers compute post-activation internally
- BackwardFunctions: re-compute pre-activation in fused linear backward
  so GELU/Swish get correct input (was passing post-activation output)
- CpuEngine: compute broadcast shape for lazy BroadcastAdd instead of
  hardcoding a._shape (fixes [3,1]+[1,4] -> [3,4])
- LazyTensorScope: mark scope as compiled when CompileInference/
  CompileTraining called to prevent double-realize on Dispose
- LazyTensorScope: restore parent scope in finally block so exceptions
  during Realize don't leave disposed scope as current
- CpuFusedOperations: hoist activation delegate lookup outside hot loops
  (was doing dictionary lookup per element)
- CpuFusedOperations: throw on unknown activation types instead of
  silently returning identity (matches ActivationRegistry.Get behavior)
- NativeInferencePool: use long multiplication to prevent int overflow
  for large buffer allocations
- CpuFusionPass: remove premature LazySource clearing on intermediates
- BackwardFunctions: fix misleading comment about beta=1 (actual: beta=0)
- Tests: add [Trait("Category","Performance")] to benchmark tests
- Tests: fix LazyGraphModeTests comment accuracy and add fusion assertion
- Tests: clarify training step benchmark comparison note

+++ src/AiDotNet.Tensors/Engines/NativeInferencePool.cs
@@ -78,7 +78,8 @@ public sealed class NativeInferencePool : IDisposable
         int key = -floatCount;
         if (!_nativeBuffers.TryGetValue(key, out var ptr))
         {
-            int byteCount = floatCount * sizeof(float);
+            // Use long multiplication to avoid int overflow for large buffers
+            long byteCount = (long)floatCount * sizeof(float);
             ptr = (IntPtr)NativeMemory.AlignedAlloc((nuint)byteCount, 64);
             _nativeBuffers[key] = ptr;
         }
@@ -94,7 +95,8 @@ public sealed class NativeInferencePool : IDisposable
         int key = doubleCount;
         if (!_nativeBuffers.TryGetValue(key, out var ptr))
         {
-            int byteCount = doubleCount * sizeof(double);
+            // Use long multiplication to avoid int overflow for large buffers
+            long byteCount = (long)doubleCount * sizeof(double);
             ptr = (IntPtr)NativeMemory.AlignedAlloc((nuint)byteCount, 64);
             _nativeBuffers[key] = ptr;
         }
